requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.88.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0

asyncpg==0.26.0
aiohttp
requests

sqlalchemy==1.4.26
sqlalchemy_utils==0.37.9

pydantic==1.8.2
starlette==0.22.0
strawberry-graphql==0.119.0
click==8.1.3
aiodataloader

https://github.com/hrbolek/uoishelpers/archive/refs/heads/main.zip
